WIP - making ZoomCapture a stateless component
